refactor(controller): split InferenceIdentityBinding controller by concern
Move the monolithic InferenceIdentityBinding controller into concern-focused files inside the same package without changing behavior.

Refs #81
